Iot/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
mysql-connector-python>=8.0.0 